requirements.txt
crewai
openai
langchain
crewai-tools
duckduckgo-search
yfinance
python-dotenv
pandas
numpy
matplotlib
crewai_tools